+++ examples/resources/iosxe_vlan_configuration/resource.tf
@@ -1,5 +1,3 @@
 resource "iosxe_vlan_configuration" "example" {
-  vlan_id           = 123
-  evpn_instance     = 123
-  evpn_instance_vni = 10123
+  vlan_id = "123"
 }